requirements.txt
streamlit>=1.40.0
plotly>=5.18.0
xlsxwriter>=3.2.0
fpdf2>=2.7.0
google-generativeai>=0.3.0
groq>=0.4.0
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.0.0
requests>=2.31.0
supabase>=2.3.0

